move assets into shared
