+++ apps/web/src/components/Builders/Landing/Customers/index.tsx
@@ -4,30 +4,30 @@ import Title from 'apps/web/src/components/base-org/typography/Title';
 import { TitleLevel } from 'apps/web/src/components/base-org/typography/Title/types';
 import Image, { StaticImageData } from 'next/image';
 
-import b3 from 'apps/web/src/components/Builders/Landing/Customers/assets/appchains/b3.svg';
-import blocklords from 'apps/web/src/components/Builders/Landing/Customers/assets/appchains/blocklords.svg';
-import illuvium from 'apps/web/src/components/Builders/Landing/Customers/assets/appchains/illuvium.svg';
-import metacade from 'apps/web/src/components/Builders/Landing/Customers/assets/appchains/metacade.png';
-import proofworks from 'apps/web/src/components/Builders/Landing/Customers/assets/appchains/proofworks.svg';
-import superchamps from 'apps/web/src/components/Builders/Landing/Customers/assets/appchains/superchamps.svg';
-import aerodrome from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/aerodrome.svg';
-import cattown from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/cattown.svg';
-import cooprecords from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/cooprecords.svg';
-import frenpet from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/frenpet.svg';
-import guild from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/guild.svg';
-import heyelsa from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/heyelsa.svg';
-import kyberswap from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/kyberswap.svg';
-import layer3 from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/layer3.svg';
-import magiceden from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/magiceden.svg';
-import moshicam from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/moshicam.svg';
-import moonwell from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/moonwell.svg';
-import opensea from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/opensea.svg';
-import pancakeswap from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/pancakeswap.svg';
-import seamless from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/seamless.svg';
-import tokeimon from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/tokeimon.svg';
-import uniswap from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/uniswap.svg';
-import virtual from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/virtual.svg';
-import zora from 'apps/web/src/components/Builders/Landing/Customers/assets/smart-wallet/zora.svg';
+import b3 from 'apps/web/src/components/Builders/Shared/assets/appchains/b3.svg';
+import blocklords from 'apps/web/src/components/Builders/Shared/assets/appchains/blocklords.svg';
+import illuvium from 'apps/web/src/components/Builders/Shared/assets/appchains/illuvium.svg';
+import metacade from 'apps/web/src/components/Builders/Shared/assets/appchains/metacade.png';
+import proofworks from 'apps/web/src/components/Builders/Shared/assets/appchains/proofworks.svg';
+import superchamps from 'apps/web/src/components/Builders/Shared/assets/appchains/superchamps.svg';
+import aerodrome from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/aerodrome.svg';
+import cattown from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/cattown.svg';
+import cooprecords from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/cooprecords.svg';
+import frenpet from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/frenpet.svg';
+import guild from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/guild.svg';
+import heyelsa from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/heyelsa.svg';
+import kyberswap from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/kyberswap.svg';
+import layer3 from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/layer3.svg';
+import magiceden from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/magiceden.svg';
+import moshicam from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/moshicam.svg';
+import moonwell from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/moonwell.svg';
+import opensea from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/opensea.svg';
+import pancakeswap from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/pancakeswap.svg';
+import seamless from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/seamless.svg';
+import tokeimon from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/tokeimon.svg';
+import uniswap from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/uniswap.svg';
+import virtual from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/virtual.svg';
+import zora from 'apps/web/src/components/Builders/Shared/assets/smart-wallet/zora.svg';
 import Link from 'apps/web/src/components/Link';
 
 import { Marquee } from 'apps/web/src/components/Builders/Shared/Marquee';
