+++ apps/web/src/components/Builders/Landing/Tools/index.tsx
@@ -2,12 +2,12 @@ import Title from 'apps/web/src/components/base-org/typography/Title';
 import { TitleLevel } from 'apps/web/src/components/base-org/typography/Title/types';
 import Image, { type StaticImageData } from 'next/image';
 import Link from 'next/link';
-import agentKit from 'apps/web/src/components/Builders/Landing/Tools/assets/agentKit.svg';
-import appchains from 'apps/web/src/components/Builders/Landing/Tools/assets/appchains.svg';
-import miniKit from 'apps/web/src/components/Builders/Landing/Tools/assets/miniKit.svg';
-import smartWallet from 'apps/web/src/components/Builders/Landing/Tools/assets/smartWallet.svg';
-import onchainKit from 'apps/web/src/components/Builders/Landing/Tools/assets/onchainKit.svg';
-import verification from 'apps/web/src/components/Builders/Landing/Tools/assets/verification.svg';
+import agentKit from 'apps/web/src/components/Builders/Shared/assets/Tools/agentKit.svg';
+import appchains from 'apps/web/src/components/Builders/Shared/assets/Tools/appchains.svg';
+import miniKit from 'apps/web/src/components/Builders/Shared/assets/Tools/miniKit.svg';
+import smartWallet from 'apps/web/src/components/Builders/Shared/assets/Tools/smartWallet.svg';
+import onchainKit from 'apps/web/src/components/Builders/Shared/assets/Tools/onchainKit.svg';
+import verification from 'apps/web/src/components/Builders/Shared/assets/Tools/verification.svg';
 
 type ToolCardProps = {
   title: string;
@@ -25,7 +25,7 @@ export function Tools() {
       <Title level={TitleLevel.Title3} className="mb-6 mt-32 font-medium md:hidden">
         The easiest and most rewarding way to build world-class onchain products.
       </Title>
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-3 md:gap-4">
+      <div className="grid grid-cols-1 gap-3 md:grid-cols-2 md:gap-4">
         <ToolCard
           title="AgentKit"
           description="Build and deploy AI agents that can interact with blockchain data and smart contracts."
@@ -73,7 +73,7 @@ function ToolCard({ title, description, icon, href }: ToolCardProps) {
       href={href}
       className="flex cursor-pointer flex-col gap-4 rounded-xl bg-dark-palette-backgroundAlternate p-6 transition-all duration-200 hover:bg-dark-gray-10"
     >
-      <div className="flex flex-col md:flex-row md:items-center gap-8">
+      <div className="flex flex-col gap-8 md:flex-row md:items-center">
         <Image src={icon} alt={title} width={64} height={64} className="h-16 w-16" />
         <div className="flex flex-col gap-2">
           <Title level={TitleLevel.Title3} as="h3" className="font-medium text-white">
