+++ apps/web/src/components/base-org/shared/TopNavigation/BuildersDropdown.tsx
@@ -2,12 +2,12 @@ import AnalyticsProvider from 'apps/web/contexts/Analytics';
 import Title from 'apps/web/src/components/base-org/typography/Title';
 import { TitleLevel } from 'apps/web/src/components/base-org/typography/Title/types';
 import { Icon } from 'apps/web/src/components/Icon/Icon';
-import agentKit from 'apps/web/src/components/Builders/Landing/Tools/assets/agentKit.svg';
-import baseNet from 'apps/web/src/components/Builders/Landing/Tools/assets/appchains.svg';
-import miniKit from 'apps/web/src/components/Builders/Landing/Tools/assets/miniKit.svg';
-import baseWallet from 'apps/web/src/components/Builders/Landing/Tools/assets/smartWallet.svg';
-import onchainKit from 'apps/web/src/components/Builders/Landing/Tools/assets/onchainKit.svg';
-import verification from 'apps/web/src/components/Builders/Landing/Tools/assets/verification.svg';
+import agentKit from 'apps/web/src/components/Builders/Shared/assets/Tools/agentKit.svg';
+import baseNet from 'apps/web/src/components/Builders/Shared/assets/Tools/appchains.svg';
+import miniKit from 'apps/web/src/components/Builders/Shared/assets/Tools/miniKit.svg';
+import baseWallet from 'apps/web/src/components/Builders/Shared/assets/Tools/smartWallet.svg';
+import onchainKit from 'apps/web/src/components/Builders/Shared/assets/Tools/onchainKit.svg';
+import verification from 'apps/web/src/components/Builders/Shared/assets/Tools/verification.svg';
 import Image, { StaticImageData } from 'next/image';
 import options from 'apps/web/src/components/base-org/shared/TopNavigation/assets/developers/options.svg';
 import bug from 'apps/web/src/components/base-org/shared/TopNavigation/assets/developers/bug.svg';
